skipping the snapshots attached with AMIs

+++ python_code/AWS/delete-snapshot-30-days.py
@@ -12,13 +12,16 @@ def lambda_handler(event, context):
 
   current_time = datetime.datetime.now()
   time_now_epoc = datetime1.strptime(str(current_time)[:19], '%Y-%m-%d %H:%M:%S')
-  logger.info(time_now_epoc)
+
   for snap in ec2_client.describe_snapshots(OwnerIds=['self'])['Snapshots']:
+
+    if not (str(snap['Description']).startswith("Created by CreateImage") or \
+      str(snap['Description']).startswith("Copied for DestinationAmi")):
+      logger.info(snap['SnapshotId'])
       start_time_epoc = datetime1.strptime(str(snap['StartTime'])[:19], '%Y-%m-%d %H:%M:%S')
-      logger.info(start_time_epoc)
       no_of_days = (time_now_epoc - start_time_epoc).days
-      logger.info(no_of_days)
       if no_of_days > 30:
+        logger.info("deleting")
         ec2_client.delete_snapshot(SnapshotId=str(snap['SnapshotId']))
       else:
-        logger.info("snapshot is less than 30 days older")
+        logger.info("skipping")
